+++ src/Modules/CMakeLists.txt
@@ -38,3 +38,4 @@ ADD_SUBDIRECTORY(Visualization)
 ADD_SUBDIRECTORY(Legacy)
 ADD_SUBDIRECTORY(BrainStimulator)
 ADD_SUBDIRECTORY(Python)
+ADD_SUBDIRECTORY(ParticleInCell)
